requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
flask>=3.0.3
pillow>=10.4.0
numpy>=1.26.4

# ── PyTorch — install MANUALLY first via https://pytorch.org/get-started/locally/
# Example (CUDA 12.1):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# Example (CPU only):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# Do NOT let pip auto-install torch — it will pick the wrong CUDA version.

# ── TripoSR ───────────────────────────────────────────────────────────────────
# Clone the repo first, then install its requirements:
#   git clone https://github.com/VAST-AI-Research/TripoSR
#   cd TripoSR && pip install -r requirements.txt
#
# The tsr package in that repo is what app.py imports as `from tsr.system import TSR`

# ── Mesh export ───────────────────────────────────────────────────────────────
trimesh>=4.4.0
# Optional but recommended for GLB/texture support:
# pip install pyembree      (faster ray casting)
# pip install pyglet        (mesh preview window)

# ── Background removal (optional but recommended) ─────────────────────────────
# pip install rembg          (uses U2Net — clean object isolation)

# ── Server ────────────────────────────────────────────────────────────────────
setuptools>=70.0.0           # Required by torchmcubes
huggingface_hub>=0.23.0      # Model weights download
